backend/requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
PyMuPDF==1.22.5
Pillow==9.5.0
